kisanhub/metoffice/requirement.txt
django-adaptor
lxml
graphos